refactor(sidebar): collapse the git root enumeration file boundary
Move the root helpers unchanged into their owning module.
